[js-api] Update for resizable buffer integration (#1961)
* [js-api] Set the implementation limit for memory types

memory32 is the full 4GiB range. memory64 is at 2^37-1
so that when the page size is converted to bytes it is
less than the JS MAX_SAFE_INTEGER value.

Leave the runtime limits for memory sizes as-is.

* [js-api] Don't allow memories without max in toResizableBuffer

* [js-api] Remove unnecessary guard on maxsize

We must allow maxsize greater than 4GiB for memory64. The
impl limit for memory64 will ensure the maxsize is within
MAX_SAFE_INTEGER.

* [js-api] Fix ordering of instructions

* [js-api] Fix units

* Fix Assert in memory.grow