Make Echidna status checking tolerate transient failure (#1856)
It often happens that checking the Echinda status too quickly after request upload results in failure (to get the validation results), but checking it shortly after that sometimes results in success. So, make the status-check script retry a couple of times on failure.

The core spec is currently experiencing more persistent failure, so allow the status checker to return success even if the check fails for now.