Refactor middle end tests (#194)
This refactors the test setup for the middle end test. Instead of one huge `TEST.md` that contains the dump output of every pass, we now have expected files for each pass separately:
```
~/build/spectec/spectec/test-middlend $ ls specification.0*exp
specification.00-elab.exp                specification.02-remove-indexed-types.exp  specification.04-else.exp            specification.06-sub.exp
specification.01-typefamily-removal.exp  specification.03-totalize.exp              specification.05-sideconditions.exp
```

Run with `dune runtest` and update with `dune promote` as usual.

This makes it easier to review a PR to see which passes are affected (at least when passes near the end are changed).
It is also very useful to have these files to diff subsequent pass dumps to understand what a pass is doing:
```
~/build/spectec/spectec/test-middlend $ diff -u specification.02-remove-indexed-types.exp specification.03-totalize.exp |head
--- specification.02-remove-indexed-types.exp   2025-11-12 11:49:22.788708924 +0100
+++ specification.03-totalize.exp       2025-11-12 11:49:22.788937041 +0100
@@ -1344,27 +1344,30 @@
   def $fsize{Fnn : Fnn}(Fnn) = $size((Fnn : Fnn <: numtype))
 
 ;; ../../../../specification/wasm-3.0/1.2-syntax.types.spectec
-def $inv_isize(nat : nat) : Inn
+def $inv_isize(nat : nat) : Inn?
   ;; ../../../../specification/wasm-3.0/1.2-syntax.types.spectec
-  def $inv_isize(32) = I32_Inn
```

Unfortunately, changes to the set and order of passes are tedious (the dune file has to be updated). See the `dune` file for instructions. If this becomes too tedious [using `dynamic_include` may be an option](https://dune.readthedocs.io/en/latest/howto/rule-generation.html), but hopefully we don’t change the sequence of passes that often.