test-middlend: do not require copying pass names in dune files (#200)
This refines upon #194, which was quickly getting boring. This refines
the setup so that the `dune` file does not need to know the names and
order of passes.

Unfortunately, `dune` does not support comparing directories
(https://github.com/ocaml/dune/issues/3567https://github.com/ocaml/dune/issues/3567),
so this uses a self-updating dune include to work-around that, as
suggested in
https://dune.readthedocs.io/en/stable/reference/dune/include.html

It uses .gitattribute to set merge=binary to avoid getting conflict markers in this file.